requirements.txt
aws-cdk-lib==2.12.0
constructs>=10.0.0,<11.0.0
attrs==21.4.0
aws-cdk-lib==2.12.0
aws-cdk.cloud-assembly-schema==1.144.0
aws-cdk.core==1.144.0
aws-cdk.cx-api==1.144.0
aws-cdk.region-info==1.144.0
boto3==1.20.49
botocore==1.23.49
cattrs==1.10.0
constructs==10.0.58
jmespath==0.10.0
jsii==1.53.0
publication==0.0.3
python-dateutil==2.8.2
s3transfer==0.5.1
six==1.16.0
typing_extensions==4.0.1
urllib3==1.26.8
